requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
dj-rest-auth==2.2.5
django-rest-swagger==2.2.0
required==0.4.0